Добавил отдельный презентер для точек маршрута (event-ов)
