+++ src/main.js
@@ -1,5 +1,5 @@
-import TripsPresenter from './presenter/trips-presenter';
-import TripsModel from './model/trips-model';
+import TripsPresenter from './presenter/trip-presenter';
+import TripsModel from './model/trip-model';
 
 const tripsModel = new TripsModel();
 const tripsPresenter = new TripsPresenter({tripsModel});
