+++ src/model/trip-model.js
